server/resume/requirements.txt
Django>=5.2.2
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
social-auth-app-django>=5.4.0
python-decouple>=3.8
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
requests>=2.31.0
reportlab>=4.0.0
Pillow>=10.0.0
python-magic>=0.4.27 